refactor: share read-only detail state
